[bindings] Remove expose-stage-data.cpp
+ didn't help much with compile times
